+++ src/sound.h
@@ -21,7 +21,7 @@ extern wav_t	snd_load_wav(const char *fname);
 #define SND_FMT_22	4
 #define SND_FMT_11	8
 
-extern wav_t	snd_load_mem(int fmt, const int *buf, size_t len);
+extern wav_t	snd_load_mem(int fmt, const short *buf, size_t len);
 extern int	snd_play_mus(char *music, int ms, int loop);
 extern int	snd_playing_mus();
 extern void     snd_stop_mus(int ms);
